+++ airflow/providers/apache/spark/CHANGELOG.rst
@@ -24,6 +24,16 @@
 Changelog
 ---------
 
+4.0.0
+.....
+
+The ``spark-binary`` connection extra could be set to any binary, but with 4.0.0 version only two values
+are allowed for it ``spark-submit`` and ``spark2-submit``.
+
+The ``spark-home`` connection extra is not allowed any more - the binary should be available on the
+PATH in order to use SparkSubmitHook and SparkSubmitOperator.
+
+
 3.0.0
 .....